requirements/dev.txt
-r prod.txt
pylint==1.3.1
django-extensions==1.4.8
Werkzeug==0.9.6
